Refactor Mesh to support dynamic page size (4KB/16KB) via templates
