+++ src/libmesh.cc
@@ -7,15 +7,26 @@
 
 #include "runtime.h"
 #include "thread_local_heap.h"
+#include "runtime_impl.h"
 
 using namespace mesh;
 
 static __attribute__((constructor)) void libmesh_init() {
   mesh::real::init();
 
-  runtime().createSignalFd();
-  runtime().installSegfaultHandler();
-  runtime().initMaxMapCount();
+  const size_t pageSize = getPageSize();
+
+  if (pageSize == 4096) {
+    runtime<4096>().createSignalFd();
+    runtime<4096>().installSegfaultHandler();
+    runtime<4096>().initMaxMapCount();
+    ThreadLocalHeap<4096>::InitTLH();
+  } else {
+    runtime<16384>().createSignalFd();
+    runtime<16384>().installSegfaultHandler();
+    runtime<16384>().initMaxMapCount();
+    ThreadLocalHeap<16384>::InitTLH();
+  }
 
   char *meshPeriodStr = getenv("MESH_PERIOD_MS");
   if (meshPeriodStr) {
@@ -23,7 +34,11 @@ static __attribute__((constructor)) void libmesh_init() {
     if (period < 0) {
       period = 0;
     }
-    runtime().setMeshPeriodMs(std::chrono::milliseconds{period});
+    if (pageSize == 4096) {
+      runtime<4096>().setMeshPeriodMs(std::chrono::milliseconds{period});
+    } else {
+      runtime<16384>().setMeshPeriodMs(std::chrono::milliseconds{period});
+    }
   }
 
   char *bgThread = getenv("MESH_BACKGROUND_THREAD");
@@ -32,7 +47,11 @@ static __attribute__((constructor)) void libmesh_init() {
 
   int shouldThread = atoi(bgThread);
   if (shouldThread) {
-    runtime().startBgThread();
+    if (pageSize == 4096) {
+      runtime<4096>().startBgThread();
+    } else {
+      runtime<16384>().startBgThread();
+    }
   }
 }
 
@@ -47,93 +66,135 @@ static __attribute__((destructor)) void libmesh_fini() {
   else if (mlevel > 2)
     mlevel = 2;
 
-  runtime().heap().dumpStats(mlevel, false);
+  if (getPageSize() == 4096) {
+    runtime<4096>().heap().dumpStats(mlevel, false);
+  } else {
+    runtime<16384>().heap().dumpStats(mlevel, false);
+  }
 }
 
 namespace mesh {
+
+template <size_t PageSize>
 ATTRIBUTE_NEVER_INLINE
-static void *allocSlowpath(size_t sz) {
-  ThreadLocalHeap *localHeap = ThreadLocalHeap::GetHeap();
+void *allocSlowpath(size_t sz) {
+  ThreadLocalHeap<PageSize> *localHeap = ThreadLocalHeap<PageSize>::GetHeap();
   return localHeap->malloc(sz);
 }
 
+template <size_t PageSize>
 ATTRIBUTE_NEVER_INLINE
 __attribute__((unused))
-static void *cxxNewSlowpath(size_t sz) {
-  ThreadLocalHeap *localHeap = ThreadLocalHeap::GetHeap();
+void *cxxNewSlowpath(size_t sz) {
+  ThreadLocalHeap<PageSize> *localHeap = ThreadLocalHeap<PageSize>::GetHeap();
   return localHeap->cxxNew(sz);
 }
 
+template <size_t PageSize>
 ATTRIBUTE_NEVER_INLINE
-static void freeSlowpath(void *ptr) {
+void freeSlowpath(void *ptr) {
   // instead of instantiating a thread-local heap on free, just free
   // to the global heap directly
-  runtime().heap().free(ptr);
+  runtime<PageSize>().heap().free(ptr);
 }
 
+template <size_t PageSize>
 ATTRIBUTE_NEVER_INLINE
-static void *reallocSlowpath(void *oldPtr, size_t newSize) {
-  ThreadLocalHeap *localHeap = ThreadLocalHeap::GetHeap();
+void *reallocSlowpath(void *oldPtr, size_t newSize) {
+  ThreadLocalHeap<PageSize> *localHeap = ThreadLocalHeap<PageSize>::GetHeap();
   return localHeap->realloc(oldPtr, newSize);
 }
 
+template <size_t PageSize>
 ATTRIBUTE_NEVER_INLINE
-static void *callocSlowpath(size_t count, size_t size) {
-  ThreadLocalHeap *localHeap = ThreadLocalHeap::GetHeap();
+void *callocSlowpath(size_t count, size_t size) {
+  ThreadLocalHeap<PageSize> *localHeap = ThreadLocalHeap<PageSize>::GetHeap();
   return localHeap->calloc(count, size);
 }
 
+template <size_t PageSize>
 ATTRIBUTE_NEVER_INLINE
-static size_t usableSizeSlowpath(void *ptr) {
-  ThreadLocalHeap *localHeap = ThreadLocalHeap::GetHeap();
+size_t usableSizeSlowpath(void *ptr) {
+  ThreadLocalHeap<PageSize> *localHeap = ThreadLocalHeap<PageSize>::GetHeap();
   return localHeap->getSize(ptr);
 }
 
+template <size_t PageSize>
 ATTRIBUTE_NEVER_INLINE
-static void *memalignSlowpath(size_t alignment, size_t size) {
-  ThreadLocalHeap *localHeap = ThreadLocalHeap::GetHeap();
+void *memalignSlowpath(size_t alignment, size_t size) {
+  ThreadLocalHeap<PageSize> *localHeap = ThreadLocalHeap<PageSize>::GetHeap();
   return localHeap->memalign(alignment, size);
 }
 }  // namespace mesh
 
 extern "C" MESH_EXPORT CACHELINE_ALIGNED_FN void *mesh_malloc(size_t sz) {
-  ThreadLocalHeap *localHeap = ThreadLocalHeap::GetHeapIfPresent();
-  if (unlikely(localHeap == nullptr)) {
-    return mesh::allocSlowpath(sz);
+  if (likely(getPageSize() == 4096)) {
+    auto *localHeap = ThreadLocalHeap<4096>::GetHeapIfPresent();
+    if (unlikely(localHeap == nullptr)) {
+      return mesh::allocSlowpath<4096>(sz);
+    }
+    return localHeap->malloc(sz);
+  } else {
+    auto *localHeap = ThreadLocalHeap<16384>::GetHeapIfPresent();
+    if (unlikely(localHeap == nullptr)) {
+      return mesh::allocSlowpath<16384>(sz);
+    }
+    return localHeap->malloc(sz);
   }
-
-  return localHeap->malloc(sz);
 }
 #define xxmalloc mesh_malloc
 
 extern "C" MESH_EXPORT CACHELINE_ALIGNED_FN void mesh_free(void *ptr) {
-  ThreadLocalHeap *localHeap = ThreadLocalHeap::GetHeapIfPresent();
-  if (unlikely(localHeap == nullptr)) {
-    mesh::freeSlowpath(ptr);
-    return;
+  if (likely(getPageSize() == 4096)) {
+    auto *localHeap = ThreadLocalHeap<4096>::GetHeapIfPresent();
+    if (unlikely(localHeap == nullptr)) {
+      mesh::freeSlowpath<4096>(ptr);
+      return;
+    }
+    localHeap->free(ptr);
+  } else {
+    auto *localHeap = ThreadLocalHeap<16384>::GetHeapIfPresent();
+    if (unlikely(localHeap == nullptr)) {
+      mesh::freeSlowpath<16384>(ptr);
+      return;
+    }
+    localHeap->free(ptr);
   }
-
-  return localHeap->free(ptr);
 }
 #define xxfree mesh_free
 
 extern "C" MESH_EXPORT CACHELINE_ALIGNED_FN void mesh_sized_free(void *ptr, size_t sz) {
-  ThreadLocalHeap *localHeap = ThreadLocalHeap::GetHeapIfPresent();
-  if (unlikely(localHeap == nullptr)) {
-    mesh::freeSlowpath(ptr);
-    return;
+  if (likely(getPageSize() == 4096)) {
+    auto *localHeap = ThreadLocalHeap<4096>::GetHeapIfPresent();
+    if (unlikely(localHeap == nullptr)) {
+      mesh::freeSlowpath<4096>(ptr);
+      return;
+    }
+    localHeap->sizedFree(ptr, sz);
+  } else {
+    auto *localHeap = ThreadLocalHeap<16384>::GetHeapIfPresent();
+    if (unlikely(localHeap == nullptr)) {
+      mesh::freeSlowpath<16384>(ptr);
+      return;
+    }
+    localHeap->sizedFree(ptr, sz);
   }
-
-  return localHeap->sizedFree(ptr, sz);
 }
 
 extern "C" MESH_EXPORT CACHELINE_ALIGNED_FN void *mesh_realloc(void *oldPtr, size_t newSize) {
-  ThreadLocalHeap *localHeap = ThreadLocalHeap::GetHeapIfPresent();
-  if (unlikely(localHeap == nullptr)) {
-    return mesh::reallocSlowpath(oldPtr, newSize);
+  if (likely(getPageSize() == 4096)) {
+    auto *localHeap = ThreadLocalHeap<4096>::GetHeapIfPresent();
+    if (unlikely(localHeap == nullptr)) {
+      return mesh::reallocSlowpath<4096>(oldPtr, newSize);
+    }
+    return localHeap->realloc(oldPtr, newSize);
+  } else {
+    auto *localHeap = ThreadLocalHeap<16384>::GetHeapIfPresent();
+    if (unlikely(localHeap == nullptr)) {
+      return mesh::reallocSlowpath<16384>(oldPtr, newSize);
+    }
+    return localHeap->realloc(oldPtr, newSize);
   }
-
-  return localHeap->realloc(oldPtr, newSize);
 }
 
 #if defined(__FreeBSD__)
@@ -153,12 +214,19 @@ extern "C" MESH_EXPORT CACHELINE_ALIGNED_FN size_t mesh_malloc_usable_size(void 
 extern "C" MESH_EXPORT CACHELINE_ALIGNED_FN size_t mesh_malloc_usable_size(const void *cptr) {
   void *ptr = const_cast<void *>(cptr);
 #endif
-  ThreadLocalHeap *localHeap = ThreadLocalHeap::GetHeapIfPresent();
-  if (unlikely(localHeap == nullptr)) {
-    return mesh::usableSizeSlowpath(ptr);
+  if (likely(getPageSize() == 4096)) {
+    auto *localHeap = ThreadLocalHeap<4096>::GetHeapIfPresent();
+    if (unlikely(localHeap == nullptr)) {
+      return mesh::usableSizeSlowpath<4096>(ptr);
+    }
+    return localHeap->getSize(ptr);
+  } else {
+    auto *localHeap = ThreadLocalHeap<16384>::GetHeapIfPresent();
+    if (unlikely(localHeap == nullptr)) {
+      return mesh::usableSizeSlowpath<16384>(ptr);
+    }
+    return localHeap->getSize(ptr);
   }
-
-  return localHeap->getSize(ptr);
 }
 #define xxmalloc_usable_size mesh_malloc_usable_size
 
@@ -167,21 +235,35 @@ extern "C" MESH_EXPORT CACHELINE_ALIGNED_FN void *mesh_memalign(size_t alignment
     throw()
 #endif
 {
-  ThreadLocalHeap *localHeap = ThreadLocalHeap::GetHeapIfPresent();
-  if (unlikely(localHeap == nullptr)) {
-    return mesh::memalignSlowpath(alignment, size);
+  if (likely(getPageSize() == 4096)) {
+    auto *localHeap = ThreadLocalHeap<4096>::GetHeapIfPresent();
+    if (unlikely(localHeap == nullptr)) {
+      return mesh::memalignSlowpath<4096>(alignment, size);
+    }
+    return localHeap->memalign(alignment, size);
+  } else {
+    auto *localHeap = ThreadLocalHeap<16384>::GetHeapIfPresent();
+    if (unlikely(localHeap == nullptr)) {
+      return mesh::memalignSlowpath<16384>(alignment, size);
+    }
+    return localHeap->memalign(alignment, size);
   }
-
-  return localHeap->memalign(alignment, size);
 }
 
 extern "C" MESH_EXPORT CACHELINE_ALIGNED_FN void *mesh_calloc(size_t count, size_t size) {
-  ThreadLocalHeap *localHeap = ThreadLocalHeap::GetHeapIfPresent();
-  if (unlikely(localHeap == nullptr)) {
-    return mesh::callocSlowpath(count, size);
+  if (likely(getPageSize() == 4096)) {
+    auto *localHeap = ThreadLocalHeap<4096>::GetHeapIfPresent();
+    if (unlikely(localHeap == nullptr)) {
+      return mesh::callocSlowpath<4096>(count, size);
+    }
+    return localHeap->calloc(count, size);
+  } else {
+    auto *localHeap = ThreadLocalHeap<16384>::GetHeapIfPresent();
+    if (unlikely(localHeap == nullptr)) {
+      return mesh::callocSlowpath<16384>(count, size);
+    }
+    return localHeap->calloc(count, size);
   }
-
-  return localHeap->calloc(count, size);
 }
 
 extern "C" {
@@ -198,61 +280,105 @@ size_t MESH_EXPORT mesh_usable_size(void *ptr) {
 // structures while forking.  This is not normally invoked when
 // libmesh is dynamically linked or LD_PRELOADed into a binary.
 void MESH_EXPORT xxmalloc_lock(void) {
-  mesh::runtime().lock();
+  if (likely(getPageSize() == 4096)) {
+    mesh::runtime<4096>().lock();
+  } else {
+    mesh::runtime<16384>().lock();
+  }
 }
 
 // ensure we don't concurrently allocate/mess with internal heap data
 // structures while forking.  This is not normally invoked when
 // libmesh is dynamically linked or LD_PRELOADed into a binary.
 void MESH_EXPORT xxmalloc_unlock(void) {
-  mesh::runtime().unlock();
+  if (likely(getPageSize() == 4096)) {
+    mesh::runtime<4096>().unlock();
+  } else {
+    mesh::runtime<16384>().unlock();
+  }
 }
 
 int MESH_EXPORT sigaction(int signum, const struct sigaction *act, struct sigaction *oldact) MESH_THROW {
-  return mesh::runtime().sigaction(signum, act, oldact);
+  if (likely(getPageSize() == 4096)) {
+    return mesh::runtime<4096>().sigaction(signum, act, oldact);
+  } else {
+    return mesh::runtime<16384>().sigaction(signum, act, oldact);
+  }
 }
 
 int MESH_EXPORT sigprocmask(int how, const sigset_t *set, sigset_t *oldset) MESH_THROW {
-  return mesh::runtime().sigprocmask(how, set, oldset);
+  if (likely(getPageSize() == 4096)) {
+    return mesh::runtime<4096>().sigprocmask(how, set, oldset);
+  } else {
+    return mesh::runtime<16384>().sigprocmask(how, set, oldset);
+  }
 }
 
 // we need to wrap pthread_create and pthread_exit so that we can
 // install our segfault handler and cleanup thread-local heaps.
 int MESH_EXPORT pthread_create(pthread_t *thread, const pthread_attr_t *attr, mesh::PthreadFn startRoutine,
                                void *arg) MESH_THROW {
-  return mesh::runtime().createThread(thread, attr, startRoutine, arg);
+  if (likely(getPageSize() == 4096)) {
+    return mesh::runtime<4096>().createThread(thread, attr, startRoutine, arg);
+  } else {
+    return mesh::runtime<16384>().createThread(thread, attr, startRoutine, arg);
+  }
 }
 
 void MESH_EXPORT ATTRIBUTE_NORETURN pthread_exit(void *retval) {
-  mesh::runtime().exitThread(retval);
+  if (likely(getPageSize() == 4096)) {
+    mesh::runtime<4096>().exitThread(retval);
+  } else {
+    mesh::runtime<16384>().exitThread(retval);
+  }
 }
 
 // Same API as je_mallctl, allows a program to query stats and set
 // allocator-related options.
 int MESH_EXPORT mesh_mallctl(const char *name, void *oldp, size_t *oldlenp, void *newp, size_t newlen) {
-  return mesh::runtime().heap().mallctl(name, oldp, oldlenp, newp, newlen);
+  if (likely(getPageSize() == 4096)) {
+    return mesh::runtime<4096>().heap().mallctl(name, oldp, oldlenp, newp, newlen);
+  } else {
+    return mesh::runtime<16384>().heap().mallctl(name, oldp, oldlenp, newp, newlen);
+  }
 }
 
 #ifdef __linux__
 
 int MESH_EXPORT epoll_wait(int __epfd, struct epoll_event *__events, int __maxevents, int __timeout) {
-  return mesh::runtime().epollWait(__epfd, __events, __maxevents, __timeout);
+  if (likely(getPageSize() == 4096)) {
+    return mesh::runtime<4096>().epollWait(__epfd, __events, __maxevents, __timeout);
+  } else {
+    return mesh::runtime<16384>().epollWait(__epfd, __events, __maxevents, __timeout);
+  }
 }
 
 int MESH_EXPORT epoll_pwait(int __epfd, struct epoll_event *__events, int __maxevents, int __timeout,
                             const __sigset_t *__ss) {
-  return mesh::runtime().epollPwait(__epfd, __events, __maxevents, __timeout, __ss);
+  if (likely(getPageSize() == 4096)) {
+    return mesh::runtime<4096>().epollPwait(__epfd, __events, __maxevents, __timeout, __ss);
+  } else {
+    return mesh::runtime<16384>().epollPwait(__epfd, __events, __maxevents, __timeout, __ss);
+  }
 }
 
 #endif
 #if __linux__
 
 ssize_t MESH_EXPORT recv(int sockfd, void *buf, size_t len, int flags) {
-  return mesh::runtime().recv(sockfd, buf, len, flags);
+  if (likely(getPageSize() == 4096)) {
+    return mesh::runtime<4096>().recv(sockfd, buf, len, flags);
+  } else {
+    return mesh::runtime<16384>().recv(sockfd, buf, len, flags);
+  }
 }
 
 ssize_t MESH_EXPORT recvmsg(int sockfd, struct msghdr *msg, int flags) {
-  return mesh::runtime().recvmsg(sockfd, msg, flags);
+  if (likely(getPageSize() == 4096)) {
+    return mesh::runtime<4096>().recvmsg(sockfd, msg, flags);
+  } else {
+    return mesh::runtime<16384>().recvmsg(sockfd, msg, flags);
+  }
 }
 #endif
 }
