+++ src/shuffle_vector.h
@@ -55,11 +55,16 @@ private:
 static_assert(sizeof(Entry) == 4, "Entry should be 4 bytes (2x uint16_t)");
 }  // namespace sv
 
+template <size_t PageSize>
 class ShuffleVector {
 private:
   DISALLOW_COPY_AND_ASSIGN(ShuffleVector);
 
 public:
+  using MiniHeapT = MiniHeap<PageSize>;
+  using BitmapT = internal::Bitmap<PageSize>;
+  using RelaxedFixedBitmapT = internal::RelaxedFixedBitmap<PageSize>;
+
   ShuffleVector() : _prng(internal::seed(), internal::seed()) {
     // set initialized = false;
   }
@@ -69,7 +74,7 @@ public:
   }
 
   // post: list has the index of all bits set to 1 in it, in a random order
-  inline uint32_t ATTRIBUTE_ALWAYS_INLINE refillFrom(uint8_t mhOffset, internal::Bitmap &bitmap) {
+  inline uint32_t ATTRIBUTE_ALWAYS_INLINE refillFrom(uint8_t mhOffset, BitmapT &bitmap) {
     d_assert(_maxCount > 0);
     d_assert_msg(_maxCount <= kMaxShuffleVectorLength, "objCount? %zu <= %zu", _maxCount, kMaxShuffleVectorLength);
 
@@ -79,10 +84,10 @@ public:
 
     // d_assert(_maxCount == _attachedMiniheap->maxCount());
 
-    internal::RelaxedFixedBitmap newBitmap{static_cast<uint32_t>(_maxCount)};
+    RelaxedFixedBitmapT newBitmap{static_cast<uint32_t>(_maxCount)};
     newBitmap.setAll(_maxCount);
 
-    internal::RelaxedFixedBitmap localBits{static_cast<uint32_t>(_maxCount)};
+    RelaxedFixedBitmapT localBits{static_cast<uint32_t>(_maxCount)};
     bitmap.setAndExchangeAll(localBits.mut_bits(), newBitmap.bits());
     localBits.invert();
 
@@ -116,11 +121,11 @@ public:
     return allocCount;
   }
 
-  void ATTRIBUTE_NEVER_INLINE refillFullSlowpath(internal::Bitmap &bitmap, size_t i) {
+  void ATTRIBUTE_NEVER_INLINE refillFullSlowpath(BitmapT &bitmap, size_t i) {
     bitmap.unset(i);
   }
 
-  FixedArray<MiniHeap, kMaxMiniheapsPerShuffleVector> &miniheaps() {
+  FixedArray<MiniHeapT, kMaxMiniheapsPerShuffleVector> &miniheaps() {
     return _attachedMiniheaps;
   }
 
@@ -199,7 +204,7 @@ public:
     return val;
   }
 
-  inline void ATTRIBUTE_ALWAYS_INLINE free(MiniHeap *mh, void *ptr) {
+  inline void ATTRIBUTE_ALWAYS_INLINE free(MiniHeapT *mh, void *ptr) {
     // const auto ptrval = reinterpret_cast<uintptr_t>(ptr);
     // const size_t off = (ptrval - _start) / _objectSize;
     // const size_t off = (ptrval - _start) * _objectSizeReciprocal;
@@ -213,7 +218,7 @@ public:
     }
   }
 
-  void ATTRIBUTE_NEVER_INLINE freeFullSlowpath(MiniHeap *mh, size_t off) {
+  void ATTRIBUTE_NEVER_INLINE freeFullSlowpath(MiniHeapT *mh, size_t off) {
     mh->freeOff(off);
   }
 
@@ -257,7 +262,8 @@ public:
     _objectSizeReciprocal = 1.0 / (float)sz;
     // Cap at 1024 for bitmap limit. Shuffle vector now uses uint16_t in sv::Entry
     // so it can track up to 1024 objects (matching bitmap capacity)
-    _maxCount = min(max(getPageSize() / sz, static_cast<size_t>(kMinStringLen)), static_cast<size_t>(1024));
+    const size_t bitmapLimit = PageSize / kMinObjectSize;
+    _maxCount = min(max(getPageSize() / sz, static_cast<size_t>(kMinStringLen)), static_cast<size_t>(bitmapLimit));
     // initially, we are unattached and therefor have no capacity.
     // Setting _off to _maxCount causes isExhausted() to return true
     // so that we don't separately have to check !isAttached() in the
@@ -271,15 +277,15 @@ private:
   int16_t _maxCount{0};                                                      // 2   42
   int16_t _off{0};                                                           // 2   44
   uint32_t _objectSize{0};                                                   // 4   48
-  FixedArray<MiniHeap, kMaxMiniheapsPerShuffleVector> _attachedMiniheaps{};  // 36  128
+  FixedArray<MiniHeapT, kMaxMiniheapsPerShuffleVector> _attachedMiniheaps{};  // 36  128
   MWC _prng;                                                                 // 36  84
   float _objectSizeReciprocal{0.0};                                          // 4   88
   uint32_t _attachedOff{0};                                                  //
   sv::Entry _list[kMaxShuffleVectorLength] CACHELINE_ALIGNED;                // 512 640
 };
 
-static_assert(HL::gcd<sizeof(ShuffleVector), CACHELINE_SIZE>::value == CACHELINE_SIZE,
-              "ShuffleVector not multiple of cacheline size!");
+// static_assert(HL::gcd<sizeof(ShuffleVector), CACHELINE_SIZE>::value == CACHELINE_SIZE,
+//               "ShuffleVector not multiple of cacheline size!");
 // FIXME should fit in 640
 // static_assert(sizeof(ShuffleVector) == 704, "ShuffleVector not expected size!");
 }  // namespace mesh
