+++ src/testing/unit/alignment.cc
@@ -13,8 +13,9 @@
 
 using namespace mesh;
 
-TEST(Alignment, NaturalAlignment) {
-  auto heap = ThreadLocalHeap::GetHeap();
+template <size_t PageSize>
+void TestNaturalAlignment() {
+  auto heap = ThreadLocalHeap<PageSize>::GetHeap();
 
   void **ptrs = reinterpret_cast<void **>(calloc(256, sizeof(void *)));
   for (size_t size = 0; size < 4096; size += 4) {
@@ -38,13 +39,22 @@ TEST(Alignment, NaturalAlignment) {
     }
   }
   heap->releaseAll();
-  mesh::runtime().heap().flushAllBins();
+  mesh::runtime<PageSize>().heap().flushAllBins();
   memset(ptrs, 0, 256 * sizeof(void *));
   free(ptrs);
 }
 
-TEST(Alignment, NonOverlapping) {
-  auto heap = ThreadLocalHeap::GetHeap();
+TEST(Alignment, NaturalAlignment) {
+  if (getPageSize() == 4096) {
+    TestNaturalAlignment<4096>();
+  } else {
+    TestNaturalAlignment<16384>();
+  }
+}
+
+template <size_t PageSize>
+void TestNonOverlapping() {
+  auto heap = ThreadLocalHeap<PageSize>::GetHeap();
 
   const auto a = heap->malloc(-8);
   const auto b = heap->malloc(-8);
@@ -53,4 +63,12 @@ TEST(Alignment, NonOverlapping) {
   // Fixes #62
   ASSERT_EQ(a, nullptr);
   ASSERT_EQ(b, nullptr);
+}
+
+TEST(Alignment, NonOverlapping) {
+  if (getPageSize() == 4096) {
+    TestNonOverlapping<4096>();
+  } else {
+    TestNonOverlapping<16384>();
+  }
 }
